pyproject.toml
[project]
name = "llmmllab-runner"
version = "0.1.0"
description = "Standalone llama.cpp server manager and request proxy"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "structlog>=24.1.0",
    "rich>=13.7.0",
    "prometheus_client>=0.22.0",
    "nvsmi>=0.4.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-httpx>=0.48b0",
]

[tool.uv]
dev-dependencies = []
